% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{add_vertices}
\alias{add_vertices}
\title{Fast Version of network::add.vertices for Edgelist-formated Network}
\usage{
add_vertices(el, nv)
}
\arguments{
\item{el}{A two-column matrix of current edges (edgelist) with an attribute
variable \code{n} containing the total current network size}

\item{nv}{A integer equal to the number of nodes to add to the network
size at the given time step}
}
\description{
This function performs a simple operation of updating the
             edgelist attribute \code{n} that tracks the total network
             size implicit in an edgelist representation of the network.
}
