% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{delete_vertices}
\alias{delete_vertices}
\title{Fast Version of network::delete.vertices for Edgelist-formated Network}
\usage{
delete_vertices(el, vid)
}
\arguments{
\item{el}{A two-column matrix of current edges (edgelist) with an attribute
variable \code{n} containing the total current network size}

\item{vid}{A vector of IDs to delete from the edgelist}
}
\description{
Given a current two-column matrix of edges and a vector of IDs
             to delete from the matrix, this function first removes any rows
             of the edgelist in which the IDs are present and then permutes
             downward the index of IDs on the edgelist that were numerically
             larger than the IDs deleted.
}

