% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{networkLitemethods}
\alias{networkLitemethods}
\alias{get.vertex.attribute.networkLite}
\alias{list.vertex.attributes.networkLite}
\alias{network.edgecount.networkLite}
\alias{as.edgelist.networkLite}
\alias{mixingmatrix.networkLite}
\alias{[<-.networkLite}
\alias{print.networkLite}
\title{networkLite Methods}
\usage{
\method{get.vertex.attribute}{networkLite}(x, attrname, ...)

\method{list.vertex.attributes}{networkLite}(x)

\method{network.edgecount}{networkLite}(x, ...)

\method{as.edgelist}{networkLite}(x, output = c("matrix", "tibble"), ...)

\method{mixingmatrix}{networkLite}(object, attr, ...)

\method{[}{networkLite}(x, i, j) <- value

\method{print}{networkLite}(x, ...)
}
\arguments{
\item{x}{a \code{networkLite} object.}

\item{attrname}{the name of a vertex attribute in \code{x}.}

\item{...}{any additional arguments.}

\item{output}{Type of edgelist to output.}

\item{object}{a \code{networkLite} object}

\item{attr}{specification of a vertex attribute in \code{object} as
described in \code{\link[ergm]{node-attr}}}

\item{i, j}{Nodal indices (must be missing for networkLite method)}

\item{value}{Value to set edges to (must be FALSE for networkLite method)}
}
\description{
S3 methods for networkLite class, for generics defined in
             network package.
}
\details{
Allows use of networkLite objects in \code{ergm_model}.
}
