% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_ergm}
\alias{simulate_ergm}
\title{Fast Version of ergm::simulate.ergm for Edgelist-formatted Network}
\usage{
simulate_ergm(p, el, coef)
}
\arguments{
\item{p}{A list of network-related nodal covariates and related terms that
is produced with \code{\link{ergm_prep}}}

\item{el}{A two-column matrix of current edges (edgelist) with an attribute
variable \code{n} containing the total current network size}

\item{coef}{Vector of coefficients associated with the formation formula}
}
\description{
Fast Version of ergm::simulate.ergm for Edgelist-formatted Network
}

