% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{stergm_prep}
\alias{stergm_prep}
\title{Converts Network Object, Formation and Dissolution Formulas, Formation and
Dissolution Coefficients, and Control Settings to a Data-Thin List Format
for STERGM Resimulation}
\usage{
stergm_prep(nw, formation, dissolution, coef.form, coef.diss, constraints,
  control = tergm::control.simulate.network())
}
\arguments{
\item{nw}{An object of class \code{network}}

\item{formation}{Right-hand sided formation formula}

\item{dissolution}{Right-hand sided dissolution formula}

\item{coef.form}{Vector of coefficients associated with the formation formula}

\item{coef.diss}{Vector of coefficients associated with the dissolution formula}

\item{constraints}{Constraints for the formation model (only \code{bd})
constraints currently supported.}

\item{control}{Control settings passed to \code{tergm::control.simulate.network}}
}
\description{
Converts Network Object, Formation and Dissolution Formulas, Formation and
Dissolution Coefficients, and Control Settings to a Data-Thin List Format
for STERGM Resimulation
}

