% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateModelTerms.R
\name{updateModelTermInputs}
\alias{updateModelTermInputs}
\title{function to appropriately update model params based on ergm model terms.
contains hard-coded implementations of some of the most commonly used ergm terms called instead of InitErgmTerm.x, checkErgmTerm, etc. 
Implemented terms are edges, nodematch, nodefactor,concurrent,degree,absdiff, nodecov, nodemix.  All other terms will return errors}
\usage{
updateModelTermInputs(dat)
}
\description{
function to appropriately update model params based on ergm model terms.
contains hard-coded implementations of some of the most commonly used ergm terms called instead of InitErgmTerm.x, checkErgmTerm, etc. 
Implemented terms are edges, nodematch, nodefactor,concurrent,degree,absdiff, nodecov, nodemix.  All other terms will return errors
}

