% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateModelTerms.R
\name{get_formula_term_args_in_formula_env}
\alias{get_formula_term_args_in_formula_env}
\title{Evaluate Ergm Model Formula Terms}
\usage{
get_formula_term_args_in_formula_env(form, termIndex)
}
\arguments{
\item{form}{an ergm model formula}

\item{termIndex}{an integer index for the formula term in form to be evaluated}
}
\description{
This is a work-around for evaluating model terms in the non-standard
             tergmLite sequence, as an alternative to ergm.getModel. Computes
             a list of the arguments to the terms in the formula with offsets
             removed, evaluated in the formula calling environment returns a
             list where the first element is the term name and subsequent (named)
             elements are the argument values named by the argument names.
}

