% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{init_tergmLite}
\alias{init_tergmLite}
\title{Initializes EpiModel netsim Object for tergmLite Simulation}
\usage{
init_tergmLite(dat)
}
\arguments{
\item{dat}{A list object containing a \code{networkDynamic} object and other
initialization information passed from \code{netsim}.}
}
\description{
Initializes EpiModel netsim Object for tergmLite Simulation
}
\examples{
library("EpiModel")
nw <- network.initialize(n = 100, directed = FALSE)
formation <- ~edges
target.stats <- 50
coef.diss <- dissolution_coefs(dissolution = ~offset(edges), duration = 20)
x <- netest(nw, formation, target.stats, coef.diss, verbose = FALSE)

param <- param.net(inf.prob = 0.3)
init <- init.net(i.num = 10)
control <- control.net(type = "SI", nsteps = 100, nsims = 5, depend = TRUE)

dat <- initialize.net(x, param, init, control)

dat <- init_tergmLite(dat)

}
