% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{networkLitemethods}
\alias{networkLitemethods}
\alias{as.network.networkLite}
\alias{get.vertex.attribute.networkLite}
\alias{list.vertex.attributes.networkLite}
\title{networkLite Methods}
\usage{
\method{as.network}{networkLite}(nw, ...)

\method{get.vertex.attribute}{networkLite}(x, attrname, ...)

\method{list.vertex.attributes}{networkLite}(x)
}
\arguments{
\item{nw}{a networkLite object}

\item{...}{any additional arguments}

\item{x}{a networkLite object}

\item{attrname}{the name of a vertex attribute in \code{x}}
}
\description{
S3 methods for networkLite class, for generics defined in
             network package.
}
\details{
Allows use of networkLite objects in \code{ergm_model}.
}
