% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_network}
\alias{simulate_network}
\title{Fast Version of tergm::simulate.network for Edgelist-formatted Network}
\usage{
simulate_network(p, el, coef.form, coef.diss, save.changes = FALSE)
}
\arguments{
\item{p}{A list of network-related nodal covariates and related terms that
is produced with \code{\link{stergm_prep}}}

\item{el}{A two-column matrix of current edges (edgelist) with an attribute
variable \code{n} containing the total current network size}

\item{coef.form}{Vector of coefficients associated with the formation formula}

\item{coef.diss}{Vector of coefficients associated with the dissolution formula}

\item{save.changes}{Logical, if \code{TRUE}, saves a matrix of changed edges
as an attribute of the output edgelist matrix}
}
\description{
Fast Version of tergm::simulate.network for Edgelist-formatted Network
}

