% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateModelTerms.R
\name{updateModelTermInputs}
\alias{updateModelTermInputs}
\title{Methods for Computing and Updating ERGM/STERGM Term Inputs}
\usage{
updateModelTermInputs(dat, network = 1)
}
\arguments{
\item{dat}{EpiModel dat object tracking simulation state}

\item{network}{Numberic number of network location for multi-network simulations.}
}
\description{
Function to appropriately update model params based on ergm model
             terms when using Edgelist-formatted representations
}
\details{
Calls ergm_model.

Implemented terms are:
 \itemize{
   \item edges
   \item nodematch
   \item nodefactor
   \item concurrent (including heterogenous by attribute)
   \item degree (including heterogenous by attribute)
   \item degrange
   \item absdiff
   \item absdiffby (in the EpiModelHIV package)
   \item nodecov
   \item nodemix
   \item absdiffnodemix (in the EpiModelHIV package)
 }
 All other terms will return errors.
}
