% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateModelTerms.R
\name{updateModelTermInputs}
\alias{updateModelTermInputs}
\title{Alternate Methods for Computing Ergm Term Inputs}
\usage{
updateModelTermInputs(dat)
}
\arguments{
\item{dat}{EpiModel dat object tracking simulation state}
}
\description{
Function to appropriately update model params based on ergm model
             terms when using 'fast_edgelist' representations.
}
\details{
Contains hard-coded implementations of some of the most commonly used
ergm terms called instead of InitErgmTerm.x, checkErgmTerm, etc.

Implemented terms are:
 \itemize{
   \item edges
   \item nodematch
   \item nodefactor
   \item concurrent
   \item degree
   \item absdiff
   \item nodecov
   \item nodemix
 }  
 All other terms will return errors.
}

